Added epoch training, validation, and saving.
Created epoch training with progress bar, for epoch. With our dataset 1 epoch is approx. 1 day.

Created validation with unseen 1/10 of the dataset.

Models are saved every 10k batches in a folder with date and time of run in the save directory.
Models are also saved with cross validation error every epoch.